Navbar - separating UserOutlet, applying loader, slight code refactoring
