+++ src/app/components/navbar/navbar.module.css
@@ -31,7 +31,7 @@
   background-color: lightgray;
 }
 
-.li>.userOutlet {
+.li .userOutlet {
   width: 124px;
 }
